backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
httpx==0.26.0

# RAG & Vector Store
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Report Generation
reportlab==4.0.7

# Voice Processing
SpeechRecognition==3.10.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

pydantic[email]
numpy==1.26.4